Naox Ver0.4:Supports parallel processing
