--- webserver/static/index.css
@@ -1,3 +1,0 @@
-h1 {
-    color: red;
-}
