--- webserver/static/index.html
@@ -1,11 +1,0 @@
-<!doctype html>
-<html lang="ja">
-<head>
-  <meta charset="UTF-8">
-  <title>Naox</title>
-  <link rel="stylesheet" href="index.css">
-</head>
-<body>
-  <img alt="logo" src="logo.png">
-  <h1>Welcome to Naox!</h1>
-</body>
